+++ ledger/tasks.py
@@ -9,16 +9,20 @@ from celery import shared_task
 
 from django.db import IntegrityError
 from django.utils import timezone
+from django.utils.html import format_html
+from django.utils.translation import gettext_lazy as _
 from esi.errors import TokenExpiredError
 from esi.models import Token
 
-from allianceauth.authentication.models import EveCharacter
+from allianceauth.authentication.models import CharacterOwnership, EveCharacter
+from allianceauth.notifications import notify
 from allianceauth.services.tasks import QueueOnce
 
 from ledger.decorators import when_esi_is_available
 from ledger.hooks import get_extension_logger
 from ledger.models.characteraudit import CharacterAudit
 from ledger.models.corporationaudit import CorporationAudit
+from ledger.models.planetary import CharacterPlanetDetails
 from ledger.task_helpers.char_helpers import (
     update_character_mining,
     update_character_wallet,
@@ -154,6 +158,70 @@ def update_character(
     )
 
     return True
+
+
+# pylint: disable=unused-argument, too-many-locals
+@shared_task(bind=True, base=QueueOnce)
+def check_planetary_alarms(self, runs: int = 0):
+    all_planets = CharacterPlanetDetails.objects.all()
+    owner_ids = {}
+    warnings = {}
+
+    for planet in all_planets:
+        if planet.is_expired():
+            character_id = planet.planet.character.character.character_id
+
+            # Determine if the character_id is part of any main character's alts
+            main_id = None
+            for main, alts in owner_ids.items():
+                if character_id in alts:
+                    main_id = main
+                    break
+
+            if not main_id:
+                owner = CharacterOwnership.objects.get(
+                    character__character_id=character_id
+                )
+                main = owner.user.profile.main_character
+                alts = main.character_ownership.user.character_ownerships.all()
+
+                owner_ids[main.character_id] = alts.values_list(
+                    "character__character_id", flat=True
+                )
+
+                main_id = main.character_id
+
+            msg = _("%(charname)s on %(planetname)s") % {
+                "charname": planet.planet.character.character.character_name,
+                "planetname": planet.planet.planet.name,
+            }
+
+            if main_id not in warnings:
+                warnings[main_id] = []
+
+            warnings[main_id].append(msg)
+
+    if warnings:
+        for main_id, messages in warnings.items():
+            owner = CharacterOwnership.objects.get(character__character_id=main_id)
+            title = _("Planetary Extractor Heads Expired")
+
+            # Split messages into chunks of 50
+            for i in range(0, len(messages), 50):
+                chunk = messages[i : i + 50]
+                msg = "\n".join(chunk)
+
+                full_message = format_html(
+                    "Following Planet Extractor Heads have expired: \n{}", msg
+                )
+                notify(
+                    title=title,
+                    message=full_message,
+                    user=owner.user,
+                    level="warning",
+                )
+                runs = runs + 1
+    logger.info("Queued %s Planetary Alarms.", runs)
 
 
 # Corporation Audit - Tasks