+++ ledger/urls.py
@@ -5,7 +5,7 @@ from django.urls import path, re_path
 from ledger.api import api
 from ledger.views.character.char_audit import add_char, fetch_memberaudit
 from ledger.views.character.character_ledger import character_admin, character_ledger
-from ledger.views.character.planetary import planetary_index
+from ledger.views.character.planetary import planetary_index, switch_alarm
 from ledger.views.corporation.corp_audit import add_corp
 from ledger.views.corporation.corp_events import (
     create_event,
@@ -59,6 +59,11 @@ urlpatterns = [
     path("events/ajax/load_events", load_events, name="load_events"),
     # -- -- Planetary
     path("planetary/", planetary_index, name="planetary_index"),
+    path(
+        "planetary/switch_alarm/<int:character_id>/planet/<int:planet_id>/",
+        switch_alarm,
+        name="switch_alarm",
+    ),
     # -- API System
     re_path(r"^api/", api.urls),
 ]